+++ semomun/Solving/Models/SectionResult.swift
@@ -9,12 +9,12 @@ import Foundation
 
 struct SectionResult {
     let title: String
-    let perfectScore: Int
-    let totalScore: Int
+    let perfectScore: Double
+    let totalScore: Double
     let totalTime: Int64
     let wrongProblems: [String]
     
-    init(title: String, perfectScore: Int, totalScore: Int, totalTime: Int64, wrongProblems: [String]) {
+    init(title: String, perfectScore: Double, totalScore: Double, totalTime: Int64, wrongProblems: [String]) {
         self.title = title
         self.perfectScore = perfectScore
         self.totalScore = totalScore
